# docs/en/contracts/control-plane-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Bastion Control Plane API
  version: 1.0.0-draft
  description: >
    Frozen contract fixture for the External Control Plane and SDK feature.
    This document is the source of truth transcribed from
    `src/control_plane/dto.rs`; `tests/control_plane_fixtures.rs` checks the
    two stay in sync. Published live at `GET /v1/openapi.yaml`
    (`src/control_plane/routes.rs`).

    Phase 5 status (final phase): every route in this document is live,
    including `POST /webhook-subscriptions` and signed, retried outbound
    delivery (`src/control_plane/webhook_delivery.rs`) — see
    `docs/en/control-plane-security.md` for the full threat model: how the
    `:action` suffix actually routes (`axum`/`matchit` can't capture a
    partial path segment), why `external_ref` lives in `business_state`, the
    SSRF guard's exact guarantee (validated once, at subscription creation),
    and which of the 5 event types this deployment can actually emit
    (`task.created`/`task.status_changed`/`task.terminal` — NOT
    `attempt.completed`/`task.escalated`, which need execution-loop wiring
    no phase has touched). This document's routes are ALSO reachable as 5
    MCP tools (`create_task`/`get_task`/`list_tasks`/`steer_task`/
    `cancel_task`, `src/control_plane/mcp_tools.rs`) via the same
    `core_ops.rs` business logic — see the security doc's "New in Phase 5"
    section. `sdk/typescript/` is the one client library shipped (no Python
...
  contact:
...
servers:
...
security:
  - bastionToken: []
paths:
  /tasks:
    post:
      operationId: createTask
      summary: Create (or idempotently return) a durable Pursue task.
      security:
        - bastionToken: [tasks:create]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
          description: >
            Repeating this call with the same owner + idempotency key must
            not create duplicate execution (acceptance criterion).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created (or the existing task for a repeated idempotency key).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ScopeDenied'
        default:
          $ref: '#/components/responses/Error'
    get:
      operationId: listTasks
      summary: List tasks visible to the authenticated credential's owner (and project, if scoped).
      security:
        - bastionToken: [tasks:read]
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TaskStatus'
      responses:
        '200':
          description: A page of tasks.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ScopeDenied'
        default:
          $ref: '#/components/responses/Error'
  /tasks/{id}:
    get:
      operationId: getTask
      summary: Fetch one task's safe summary.
      security:
        - bastionToken: [tasks:read]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: The task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ScopeDenied'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /tasks/{id}/attempts:
    get:
      operationId: listTaskAttempts
      summary: Safe evidence/verdict timeline for one task.
      security:
        - bastionToken: [tasks:read]
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: A page of attempt summaries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttemptListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ScopeDenied'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /tasks/{id}:pause:
    post:
      operationId: pauseTask
      summary: Pause a task via expected revision.
      security:
        - bastionToken: [tasks:control]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevisionGuardedRequest'
      responses:
        '200':
          description: The task after transition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ScopeDenied'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/RevisionConflict'
        default:
          $ref: '#/components/responses/Error'
  /tasks/{id}:resume:
    post:
      operationId: resumeTask
      summary: Resume a paused task via expected revision.
      security:
        - bastionToken: [tasks:control]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevisionGuardedRequest'
      responses:
        '200':
          description: The task after transition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ScopeDenied'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/RevisionConflict'
        default:
          $ref: '#/components/responses/Error'
  /tasks/{id}:steer:
    post:
      operationId: steerTask
      summary: Add human/orchestrator guidance to an in-flight task.
      security:
        - bastionToken: [tasks:control]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/RevisionGuardedRequest'
                - type: object
                  required: [guidance]
                  properties:
                    guidance:
                      type: string
      responses:
        '200':
          description: The task after the steer note is recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ScopeDenied'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/RevisionConflict'
        default:
          $ref: '#/components/responses/Error'
  /tasks/{id}:cancel:
    post:
      operationId: cancelTask
      summary: Cancel a task via expected revision, with an auditable stop reason.
      security:
        - bastionToken: [tasks:control]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevisionGuardedRequest'
      responses:
        '200':
          description: The task after transition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ScopeDenied'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/RevisionConflict'
        default:
          $ref: '#/components/responses/Error'
  /webhook-subscriptions:
    post:
      operationId: createWebhookSubscription
      summary: Register a signed event target.
      security:
        - bastionToken: [webhooks:manage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookSubscriptionRequest'
      responses:
        '201':
          description: Subscription created. The HMAC signing secret is returned exactly once in this response only.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscriptionResource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ScopeDenied'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bastionToken:
      type: apiKey
      in: header
      name: x-bastion-token
      description: >
        Static, scoped integration credential (src/control_plane/credential.rs),
        format `bcp_<opaque>`. NOT `Authorization: Bearer` — this repo's
        auth-header convention is `x-bastion-token` everywhere (the webhook
        channel, the MCP server, and now here), resolved in
        docs/en/control-plane-security.md's "which header" decision.
        Not OAuth. Scopes: tasks:read, tasks:create, tasks:control, webhooks:manage.
  parameters:
    TaskId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Opaque task id — never sequential, never guessable-safe by itself (authorization is still enforced per-request).
  schemas:
    TaskMode:
      type: string
      enum: [respond, act, pursue]
      description: Mirrors bastion_runtime::task::ExecutionMode. Only "pursue" is ever caller-creatable via this API.
    TaskStatus:
      type: string
      enum: [pending, running, awaiting_approval, paused, completed, escalated, cancelled, failed]
      description: Mirrors bastion_runtime::task::TaskStatus 1:1.
...
    TaskResource:
      type: object
      required: [id, owner_id, mode, objective, status, created_at, updated_at, revision, budget_summary, attempts]
      properties:
        id: { type: string }
        owner_id: { type: string }
        external_ref: { type: [string, "null"] }
        mode:
          $ref: '#/components/schemas/TaskMode'
        objective: { type: string }
        status:
          $ref: '#/components/schemas/TaskStatus'
        stop_reason:
          oneOf:
            - $ref: '#/components/schemas/StopReason'
            - type: "null"
        created_at: { type: integer, format: int64 }
        updated_at: { type: integer, format: int64 }
        revision: { type: integer, format: int64 }
        budget_summary:
          $ref: '#/components/schemas/BudgetSummary'
        attempts:
          type: array
          items:
            $ref: '#/components/schemas/AttemptSummary'
    CreateTaskBounds:
      type: object
      properties:
        max_steps: { type: [integer, "null"], format: int32 }
        max_cost_usd: { type: [number, "null"] }
    CreateTaskRequest:
      type: object
      required: [objective]
      properties:
        objective: { type: string }
        external_ref: { type: [string, "null"] }
        acceptance:
          type: array
          items: { type: string }
        bounds:
          oneOf:
            - $ref: '#/components/schemas/CreateTaskBounds'
            - type: "null"
    TaskListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaskResource'
        next_cursor: { type: [string, "null"] }
    AttemptListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AttemptSummary'
        next_cursor: { type: [string, "null"] }
    RevisionGuardedRequest:
      type: object
      required: [expected_revision]
      properties:
        expected_revision:
          type: integer
          format: int64
          description: Optimistic-concurrency guard — mirrors TaskCase.revision. A stale value yields 409.
    WebhookSubscriptionRequest:
      type: object
      required: [target_url]
      properties:
        target_url:
          type: string
          format: uri
          description: Validated against SSRF (private/loopback ranges) by the handler in a later phase — not enforced by this schema alone.
        event_types:
          type: array
          items:
            type: string
            enum: [task.created, task.status_changed, attempt.completed, task.escalated, task.terminal]
...
    ErrorEnvelope:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
...
  responses:
    Unauthorized:
      description: Missing or unrecognized credential.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ScopeDenied:
      description: Credential authenticated but lacks the required scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: No task/resource with that id is visible to this credential's owner.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RevisionConflict:
      description: expected_revision did not match the task's current revision (optimistic-concurrency conflict).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error:
      description: Unclassified error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'